initializeFirestoreにまとめた
